+++ changelog.md
@@ -3,6 +3,7 @@ v0.0.8-alpha
 * Allow research tables to hold writing implements between uses
 * Show spell cooldowns visually
 * Record teleport distance statistic
+* Properly trigger pixie aggro when attacking with spells
 
 v0.0.7-alpha
 ------------